Move types in rhsplib to core
