+++ packages/core/src/DebugGroup.ts
@@ -1,0 +1,14 @@
+export enum DebugGroup {
+    Main = 1,
+    TransmitterToHost,
+    ReceiverFromHost,
+    ADC,
+    PWMAndServo,
+    ModuleLED,
+    DigitalIO,
+    I2C,
+    Motor0,
+    Motor1,
+    Motor2,
+    Motor3,
+}
