+++ packages/core/src/DioDirection.ts
@@ -1,0 +1,4 @@
+export enum DioDirection {
+    Input,
+    Output,
+}
