+++ packages/core/src/DiscoveredAddresses.ts
@@ -1,0 +1,5 @@
+export interface DiscoveredAddresses {
+    parentAddress: number;
+    childAddresses: number[];
+    numberOfChildModules: number;
+}
