+++ packages/core/src/I2CReadStatus.ts
@@ -1,0 +1,5 @@
+export interface I2CReadStatus {
+    i2cTransactionStatus: number;
+    numBytesRead: number;
+    bytes: number[];
+}
