+++ packages/core/src/I2CSpeedCode.ts
@@ -1,0 +1,4 @@
+export enum I2CSpeedCode {
+    SpeedCode100_Kbps,
+    SpeedCode400_Kbps,
+}
