+++ packages/core/src/I2CWriteStatus.ts
@@ -1,0 +1,4 @@
+export interface I2CWriteStatus {
+    i2cTransactionStatus: number;
+    numBytesWritten: number;
+}
