+++ packages/core/src/LedPattern.ts
@@ -1,0 +1,18 @@
+export interface LedPattern {
+    rgbtPatternStep0: number;
+    rgbtPatternStep1: number;
+    rgbtPatternStep2: number;
+    rgbtPatternStep3: number;
+    rgbtPatternStep4: number;
+    rgbtPatternStep5: number;
+    rgbtPatternStep6: number;
+    rgbtPatternStep7: number;
+    rgbtPatternStep8: number;
+    rgbtPatternStep9: number;
+    rgbtPatternStep10: number;
+    rgbtPatternStep11: number;
+    rgbtPatternStep12: number;
+    rgbtPatternStep13: number;
+    rgbtPatternStep14: number;
+    rgbtPatternStep15: number;
+}
