+++ packages/core/src/ModuleInterface.ts
@@ -1,0 +1,5 @@
+export interface ModuleInterface {
+    name: string;
+    firstPacketID: number;
+    numberIDValues: number;
+}
