+++ packages/core/src/ModuleStatus.ts
@@ -1,0 +1,4 @@
+export interface ModuleStatus {
+    statusWord: number;
+    motorAlerts: number;
+}
