+++ packages/core/src/PidCoefficients.ts
@@ -1,0 +1,5 @@
+export interface PidCoefficients {
+    p: number;
+    i: number;
+    d: number;
+}
