+++ packages/core/src/Rgb.ts
@@ -1,0 +1,5 @@
+export interface Rgb {
+    red: number;
+    green: number;
+    blue: number;
+}
