+++ packages/core/src/SerialFlowControl.ts
@@ -1,0 +1,5 @@
+export enum SerialFlowControl {
+    None = 0,
+    Hardware,
+    Software,
+}
