+++ packages/core/src/SerialParity.ts
@@ -1,0 +1,5 @@
+export enum SerialParity {
+    None = 0,
+    Odd,
+    Even,
+}
