+++ packages/core/src/VerbosityLevel.ts
@@ -1,0 +1,6 @@
+export enum VerbosityLevel {
+    Off = 0,
+    Level1,
+    Level2,
+    Level3,
+}
