+++ packages/core/src/Version.ts
@@ -1,0 +1,8 @@
+export interface Version {
+    engineeringRevision: number;
+    minorVersion: number;
+    majorVersion: number;
+    minorHwRevision: number;
+    majorHwRevision: number;
+    hwType: number;
+}
