+++ packages/rhsplib/package.json
@@ -12,6 +12,7 @@
         "url": "git+https://github.com/REVrobotics/node-rhsplib.git"
     },
     "dependencies": {
+        "@rev-robotics/rev-hub-core": "1.0.0",
         "node-addon-api": "^6.0.0",
         "node-gyp-build": "^4.6.0"
     },
